Создание дашборда по пользовательским событиям для агрегатора новостей/requirements.txt
pandas
sqlalchemy

dash
dash_core_components
dash_html_components
dash.dependencies
plotly.graph_objs

